backend/requirements.txt
# InboxGrove Backend Dependencies
# Production-grade SaaS backend with complete billing, provisioning, and monitoring

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Async & Background Tasks
celery==5.3.4
python-dateutil==2.8.2

# Payment Processing
stripe==7.4.0

# API Documentation
python-multipart==0.0.6

# Security & Authentication
pyjwt==2.8.1
bcrypt==4.1.1
cryptography==41.0.7
python-jose==3.3.0

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Monitoring & Logging
sentry-sdk==1.38.0
prometheus-client==0.19.0
python-json-logger==2.0.7

# Email
python-dotenv==1.0.0
aiosmtplib==3.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
